Add dedicated stop_stim function to Controller class
Previously, when the stimulation needed to stop, the controller's
running_stim flag was set to False. This did not change the text
of the Start/Stop button back to "Start", leading to buggy behavior.
Now, the controller's function stop_stim() is called to stop the
stimulation *and* change the text of the Start/Stop button at the
same time, much like the start_stim function.