Merge with colleenjg/Mouse-Stimulation, fix bugs, add black proj. window
There is now one repository for both Michel's and Colleen's stimulation
code. No need to have two separate ones as they are basically the same,
except for different types of stimuli.

A new addition is that when the stimulus is being shown on the monitor
rather than the projector, a black screen will be shown on the
projector, blocking the desktop wallpaper that would normally show.

Also fixed a bug with the optomotor stimulus, now the merge position
works properly. Fixed some other small bugs too.